requirements.txt
rich
psutil
requests
speedtest-cli
